Add files via upload
The Lazza Ice Cream website  showcases a responsive, Visually appealing ice cream shop landing page built using HTML, CSS and Bootstrap.The Webpage includes a navigation bar, a hero section, a menu grid, a gallery of ice cream flavours and a contact.  The layout is fully responsive to ensure it looks great on all screen sizes, and it is designed to both user friendly and virtual attractive.